// src/models/ride/ride-list.ts
const rideListResponse = createApiResponseSchema(rideListResponseSchema);

export type RideListResponse = z.infer<typeof rideListResponse>;
export type RideList = RideListResponse["data"];
// ...
const rideListFilterSchema = z.object({
    page: z.number(),
    limit: z.number(),
    start_date_time: z.string().optional(),
    end_date_time: z.string().optional(),
    ride_status: z.array(z.string()).optional(),
    search_driver: z.string().optional(),
    search_hitcher: z.string().optional(),
    search_route: z.string().optional(),
    search_vehicle: z.string().optional(),
});

export type RideListFilter = z.infer<typeof rideListFilterSchema>;